+++ packages/clerk-js/src/core/modules/checkout/instance.ts
@@ -24,7 +24,7 @@ function createCheckoutInstance(
 ): __experimental_CheckoutInstance {
   const { for: forOrganization, planId, planPeriod } = options;
 
-  if (!clerk.user) {
+  if (!clerk.isSignedIn || !clerk.user) {
     throw new Error('Clerk: User is not authenticated');
   }