+++ packages/clerk-js/src/ui/components/SessionTasks/index.tsx
@@ -1,16 +1,20 @@
 import { useClerk } from '@clerk/shared/react';
 import { eventComponentMounted } from '@clerk/shared/telemetry';
-import { useCallback, useContext, useEffect, useRef, useState } from 'react';
+import { useContext, useEffect, useRef } from 'react';
 
 import { Card } from '@/ui/elements/Card';
 import { withCardStateProvider } from '@/ui/elements/contexts';
 import { LoadingCardContainer } from '@/ui/elements/LoadingCard';
 
-import { SESSION_TASK_ROUTE_BY_KEY } from '../../../core/sessionTasks';
+import { INTERNAL_SESSION_TASK_ROUTE_BY_KEY } from '../../../core/sessionTasks';
 import { SignInContext, SignUpContext } from '../../../ui/contexts';
-import { SessionTasksContext, useSessionTasksContext } from '../../contexts/components/SessionTasks';
+import {
+  SessionTasksContext,
+  TaskSelectOrganizationContext,
+  useSessionTasksContext,
+} from '../../contexts/components/SessionTasks';
 import { Route, Switch, useRouter } from '../../router';
-import { ForceOrganizationSelectionTask } from './tasks/ForceOrganizationSelection';
+import { TaskSelectOrganization } from './tasks/TaskSelectOrganization';
 
 const SessionTasksStart = () => {
   const clerk = useClerk();
@@ -36,10 +40,16 @@ const SessionTasksStart = () => {
 };
 
 function SessionTaskRoutes(): JSX.Element {
+  const ctx = useSessionTasksContext();
+
   return (
     <Switch>
-      <Route path={SESSION_TASK_ROUTE_BY_KEY['org']}>
-        <ForceOrganizationSelectionTask />
+      <Route path={INTERNAL_SESSION_TASK_ROUTE_BY_KEY['select-organization']}>
+        <TaskSelectOrganizationContext.Provider
+          value={{ componentName: 'TaskSelectOrganization', redirectUrlComplete: ctx.redirectUrlComplete }}
+        >
+          <TaskSelectOrganization />
+        </TaskSelectOrganizationContext.Provider>
       </Route>
       <Route index>
         <SessionTasksStart />
@@ -56,7 +66,6 @@ export const SessionTask = withCardStateProvider(() => {
   const { navigate } = useRouter();
   const signInContext = useContext(SignInContext);
   const signUpContext = useContext(SignUpContext);
-  const [isNavigatingToTask, setIsNavigatingToTask] = useState(false);
   const currentTaskContainer = useRef<HTMLDivElement>(null);
 
   const redirectUrlComplete =
@@ -67,10 +76,6 @@ export const SessionTask = withCardStateProvider(() => {
   // for example by using browser back navigation. Since there are no pending tasks,
   // we redirect them to their intended destination.
   useEffect(() => {
-    if (isNavigatingToTask) {
-      return;
-    }
-
     // Tasks can only exist on pending sessions, but we check both conditions
     // here to be defensive and ensure proper redirection
     const task = clerk.session?.currentTask;
@@ -80,14 +85,7 @@ export const SessionTask = withCardStateProvider(() => {
     }
 
     clerk.telemetry?.record(eventComponentMounted('SessionTask', { task: task.key }));
-  }, [clerk, navigate, isNavigatingToTask, redirectUrlComplete]);
-
-  const nextTask = useCallback(() => {
-    setIsNavigatingToTask(true);
-    return clerk
-      .__internal_navigateToTaskIfAvailable({ redirectUrlComplete })
-      .finally(() => setIsNavigatingToTask(false));
-  }, [clerk, redirectUrlComplete]);
+  }, [clerk, navigate, redirectUrlComplete]);
 
   if (!clerk.session?.currentTask) {
     return (
@@ -105,7 +103,7 @@ export const SessionTask = withCardStateProvider(() => {
   }
 
   return (
-    <SessionTasksContext.Provider value={{ nextTask, redirectUrlComplete, currentTaskContainer }}>
+    <SessionTasksContext.Provider value={{ redirectUrlComplete, currentTaskContainer }}>
       <SessionTaskRoutes />
     </SessionTasksContext.Provider>
   );