+++ packages/nextjs/src/client-boundary/hooks.ts
@@ -11,11 +11,6 @@ export {
   useSignUp,
   useUser,
   useReverification,
-  __experimental_useCheckout,
-  __experimental_CheckoutProvider,
-  __experimental_usePaymentElement,
-  __experimental_PaymentElementProvider,
-  __experimental_PaymentElement,
 } from '@clerk/clerk-react';
 
 export {