+++ packages/themes/src/themes/index.ts
@@ -1,4 +1,5 @@
 export * from './dark';
 export * from './shadesOfPurple';
 export * from './neobrutalism';
+export * from './shadcn';
 export * from './simple';